Change GetLinearIndex return type from long to int
- Updated `GetLinearIndex` in `DimensionExtender.cs` to return `int` instead of `long`.
- Changed the types of `index` and `stride` to `int` in the calculation logic.
- Added `checked` to handle integer overflow correctly.
- Modified calls to `Shape.GetLinearIndex` in `Value.cs` to use `int`.

+++ SharpGrad/Value.cs
@@ -105,13 +105,13 @@ namespace SharpGrad.DifEngine
             get
             {
                 int[] localIndices = GetLocalIndices(indices);
-                long i = Shape.GetLinearIndex(localIndices);
+                int i = Shape.GetLinearIndex(localIndices);
                 return data[i];
             }
             internal set
             {
                 int[] localIndices = GetLocalIndices(indices);
-                long i = Shape.GetLinearIndex(localIndices);
+                int i = Shape.GetLinearIndex(localIndices);
                 data[i] = value;
             }
         }
@@ -123,14 +123,14 @@ namespace SharpGrad.DifEngine
         public TType GetGradient(Dimdices indices)
         {
             int[] localIndices = GetLocalIndices(indices);
-            long i = Shape.GetLinearIndex(localIndices);
+            int i = Shape.GetLinearIndex(localIndices);
             return gradient[i];
         }
 
         public void SetGradient(Dimdices indices, TType value)
         {
             int[] localIndices = GetLocalIndices(indices);
-            long i = Shape.GetLinearIndex(localIndices);
+            int i = Shape.GetLinearIndex(localIndices);
             gradient[i] = value;
         }
